+++ NEWS.md
@@ -1,7 +1,7 @@
 
 
 # archeofrag.gui 1.2.0
-Released: 2026-09-XX
+Released: 2026-09-16
 
 * Use the new function `archeofrag::frag.cohesion.ranking()` instead of in-built similar code.
 * Exploration of the full graph, with no selection of spatial variable, is now possible and set as the default setting.
@@ -9,6 +9,7 @@ Released: 2026-09-XX
 * New internal `.check_weighting_variable()` functions.
 * New 'Datasets' tab displaying summary information about 'archeofrag' embedded datasets.
 * Significantly improved 'Spatial units optimisation' performances.
+* Data export to the 'fabryka' application for spatial orientation analysis. 
 
 # archeofrag.gui 1.1.4
 Released: 2026-01-08